+++ Holovibes/sources/thread/file_frame_read_worker.cc
@@ -96,7 +96,7 @@ bool FileFrameReadWorker::init_frame_buffers()
         if (cleanupCpu)
             cudaXFreeHost(cpu_frame_buffer_);
         if (cleanupGpu)
-            cudaXFree(gpu_frame_buffer_);
+            cudaXFree(gpu_file_frame_buffer_);
 
         return false; // Explicitly return false for clarity.
     };
@@ -109,7 +109,7 @@ bool FileFrameReadWorker::init_frame_buffers()
         return handleError("Not enough CPU RAM to read file");
     }
 
-    error_code = cudaXRMalloc(&gpu_frame_buffer_, buffer_size);
+    error_code = cudaXRMalloc(&gpu_file_frame_buffer_, buffer_size);
     if (error_code != cudaSuccess)
     {
         return handleError("Not enough GPU DRAM to read file", true);
@@ -127,7 +127,7 @@ bool FileFrameReadWorker::init_frame_buffers()
 void FileFrameReadWorker::free_frame_buffers()
 {
     cudaXFree(gpu_packed_buffer_);
-    cudaXFree(gpu_frame_buffer_);
+    cudaXFree(gpu_file_frame_buffer_);
     cudaXFreeHost(cpu_frame_buffer_);
 }
 
@@ -228,13 +228,13 @@ size_t FileFrameReadWorker::read_copy_file(size_t frames_to_read)
 
                 // Convert 12bit frame to 16bit
                 if (flag_packed == 12)
-                    unpack_12_to_16bit((short*)(gpu_frame_buffer_ + i * frame_size_),
+                    unpack_12_to_16bit((short*)(gpu_file_frame_buffer_ + i * frame_size_),
                                        frame_size_ / 2,
                                        (unsigned char*)gpu_packed_buffer_,
                                        packed_frame_size,
                                        stream_);
                 else if (flag_packed == 10)
-                    unpack_10_to_16bit((short*)(gpu_frame_buffer_ + i * frame_size_),
+                    unpack_10_to_16bit((short*)(gpu_file_frame_buffer_ + i * frame_size_),
                                        frame_size_ / 2,
                                        (unsigned char*)gpu_packed_buffer_,
                                        packed_frame_size,
@@ -244,8 +244,12 @@ size_t FileFrameReadWorker::read_copy_file(size_t frames_to_read)
         else
         {
             // Memcopy in the gpu buffer
-            cudaXMemcpyAsync(gpu_frame_buffer_, cpu_frame_buffer_, frames_total_size, cudaMemcpyHostToDevice, stream_);
-            // cudaXMemcpy(gpu_frame_buffer_, cpu_frame_buffer_, frames_total_size, cudaMemcpyHostToDevice);
+            cudaXMemcpyAsync(gpu_file_frame_buffer_,
+                             cpu_frame_buffer_,
+                             frames_total_size,
+                             cudaMemcpyHostToDevice,
+                             stream_);
+            // cudaXMemcpy(gpu_file_frame_buffer_, cpu_frame_buffer_, frames_total_size, cudaMemcpyHostToDevice);
         }
 
         cudaStreamSynchronize(stream_);
@@ -269,8 +273,7 @@ void FileFrameReadWorker::enqueue_loop(size_t nb_frames_to_enqueue)
 
         if (Holovibes::instance().is_cli)
         {
-            while (api::get_input_queue()->get_size() ==
-                       api::get_input_queue()->get_total_nb_frames() &&
+            while (api::get_input_queue()->get_size() == api::get_input_queue()->get_total_nb_frames() &&
                    !stop_requested_)
             {
             }
@@ -279,7 +282,7 @@ void FileFrameReadWorker::enqueue_loop(size_t nb_frames_to_enqueue)
         if (stop_requested_)
             break;
 
-        input_queue_.load()->enqueue(gpu_frame_buffer_ + frames_enqueued * frame_size_,
+        input_queue_.load()->enqueue(gpu_file_frame_buffer_ + frames_enqueued * frame_size_,
                                      api::get_input_queue_location() == holovibes::Device::GPU
                                          ? cudaMemcpyDeviceToDevice
                                          : cudaMemcpyDeviceToHost);
@@ -292,7 +295,7 @@ void FileFrameReadWorker::enqueue_loop(size_t nb_frames_to_enqueue)
     }
 
     // Synchronize forced, because of the cudaMemcpyAsync we have to finish to
-    // enqueue the gpu_frame_buffer_ before storing next read frames in it.
+    // enqueue the gpu_file_frame_buffer_ before storing next read frames in it.
     //
     // With load_file_in_gpu_ == true, all the file in in the buffer,
     // so we don't have to sync