+++ CHANGELOG.md
@@ -553,7 +553,7 @@
 - Filter2D: rework filter2D
 - CLI: add progress bar and other info
 - Fix: minor gui bugfixes
-- Fix: batch_input_queue async copy and input file gpu_frame_buffer synchro
+- Fix: batch_input_queue async copy and input file gpu_file_frame_buffer synchro
 
 ### 9.1